+++ tests/testthat/test-MaestroPipeline.R
@@ -7,14 +7,12 @@ test_that("Simple pipeline, no errors", {
   pipeline <- pipeline_list$MaestroPipelines[[1]]
   expect_s3_class(pipeline, "MaestroPipeline")
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status())
   expect_null(pipeline$get_outputs())
   expect_null(pipeline$get_inputs())
 
   pipeline$run(quiet = TRUE)
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status()[c("invoked", "success", "errors", "warnings", "messages")])
 })
 
@@ -25,12 +23,10 @@ test_that("Pipeline with warnings", {
   )
   pipeline <- pipeline_list$MaestroPipelines[[1]]
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status())
 
   pipeline$run(quiet = TRUE)
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status()[c("invoked", "success", "errors", "warnings", "messages")])
 })
 
@@ -43,7 +39,6 @@ test_that("Pipeline with errors", {
 
   expect_error(pipeline$run(quiet = TRUE))
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status()[c("invoked", "success", "errors", "warnings", "messages")])
 })
 
@@ -55,14 +50,12 @@ test_that("Pipeline with arguments are correctly passed", {
 
   pipeline <- pipeline_list$MaestroPipelines[[1]]
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status()[c("invoked", "success", "errors", "warnings", "messages")])
 
   pipeline$run(resources = list(
     vals = 1:5
   ), quiet = TRUE)
 
-  expect_snapshot(pipeline$get_schedule())
   expect_snapshot(pipeline$get_status()[c("invoked", "success", "errors", "warnings", "messages")])
 })