+++ tests/testthat/test_pipelines/test_pipeline_days_bad3.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency hourly
+#' @maestroDays 50 60
+some_days <- function() {
+  invisible()
+}